// UAAEcommerce/ClientApp/src/app/models/user.model.ts
export class UserModel {
  id: string = null;
  email: string = null;
  firstName: string = null;
  lastName: string = null;
  username: string = null;
  phoneNumber: string = null;
  address: string = null;
  roles: Array<string> = [];
  fbUserId: string = null;
  googleUserId: string = null;
  profilePicture: string = null;
  tipoDocumento: string = null;
  date:string= null;
  city:string=null;
  documento: string = null;
  ruc: string = null;
  razonSocial: string = null;
}